Break out Refbox, TeamInfo, and VisionDetection into separate source and header files
