+++ src/software/sensor_fusion/refbox_data.h
@@ -1,10 +1,10 @@
 #pragma once
 
-#include <ostream>
 #include <vector>
 
 #include "software/new_geom/point.h"
 #include "software/proto/ssl_referee.pb.h"
+#include "software/sensor_fusion/team_info.h"
 #include "software/time/timestamp.h"
 #include "software/util/printable_enum_macro/printable_enum_macro.h"
 
@@ -70,214 +70,6 @@ MAKE_ENUM(RefboxStage,
           // The game is over.
           POST_GAME, REFBOX_STAGE_COUNT);
 
-class TeamInfo
-{
-   public:
-    /**
-     * Default constructor
-     */
-    TeamInfo();
-
-    /**
-     * Constructor that initializes all fields
-     *
-     * @param name name
-     * @param score score
-     * @param red_cards red_cards
-     * @param yellow_card_times yellow_card_times
-     * @param yellow_cards yellow_cards
-     * @param timeouts timeouts
-     * @param timeout_time timeout_time
-     * @param goalkeeper goalkeeper
-     * @param foul_counter foul_counter
-     * @param ball_placement_failures ball_placement_failures
-     * @param can_place_ball can_place_ball
-     * @param max_allowed_bots max_allowed_bots
-     */
-    TeamInfo(std::string name, int score, int red_cards,
-             std::vector<int> yellow_card_times, int yellow_cards, int timeouts,
-             int timeout_time, int goalkeeper, int foul_counter,
-             int ball_placement_failures, bool can_place_ball, int max_allowed_bots)
-        : name(name),
-          score(score),
-          red_cards(red_cards),
-          yellow_card_times(yellow_card_times),
-          yellow_cards(yellow_cards),
-          timeouts(timeouts),
-          timeout_time(timeout_time),
-          goalkeeper(goalkeeper),
-          foul_counter(foul_counter),
-          ball_placement_failures(ball_placement_failures),
-          can_place_ball(can_place_ball),
-          max_allowed_bots(max_allowed_bots)
-    {
-    }
-
-    /**
-     * Return name
-     *
-     * @return name
-     */
-    std::string getName(void) const
-    {
-        return name;
-    }
-
-    /**
-     * Return score
-     *
-     * @return score
-     */
-    int getScore(void) const
-    {
-        return score;
-    }
-
-    /**
-     * Return number of red cards
-     *
-     * @return number of red cards
-     */
-    int getRedCards(void) const
-    {
-        return red_cards;
-    }
-
-    /**
-     * Return yellow card times
-     *
-     * @return yellow card times
-     */
-    std::vector<int> getYellowCardTimes(void) const
-    {
-        return yellow_card_times;
-    }
-
-    /**
-     * Return number of red cards
-     *
-     * @return number of red cards
-     */
-    int getYellowCards(void) const
-    {
-        return yellow_cards;
-    }
-
-    /**
-     * Return number of timeouts
-     *
-     * @return number of timeouts
-     */
-    int getTimeouts(void) const
-    {
-        return timeouts;
-    }
-
-    /**
-     * Return timeout time
-     *
-     * @return timeout time
-     */
-    int getTimeoutTime(void) const
-    {
-        return timeout_time;
-    }
-
-    /**
-     * Return goalkeeper
-     *
-     * @return goalkeeper
-     */
-    int getGoalkeeper(void) const
-    {
-        return goalkeeper;
-    }
-
-    /**
-     * Return foul counter
-     *
-     * @return foul counter
-     */
-    int getFoulCounter(void) const
-    {
-        return foul_counter;
-    }
-
-    /**
-     * Return ball placement failures
-     *
-     * @return ball placement failures
-     */
-    int getBallPlacementFailures(void) const
-    {
-        return ball_placement_failures;
-    }
-
-    /**
-     * Return can place ball
-     *
-     * @return can place ball
-     */
-    bool getCanPlaceBall(void) const
-    {
-        return can_place_ball;
-    }
-
-    /**
-     * Return max allowed bots
-     *
-     * @return max allowed bots
-     */
-    int getMaxAllowedBots(void) const
-    {
-        return max_allowed_bots;
-    }
-
-    bool operator==(const TeamInfo &other) const
-    {
-        return ((name == other.getName()) && (score == other.getScore()) &&
-                (red_cards == other.getRedCards()) &&
-                (yellow_card_times == other.getYellowCardTimes()) &&
-                (yellow_cards == other.getYellowCards()) &&
-                (timeouts == other.getTimeouts()) &&
-                (timeout_time == other.getTimeoutTime()) &&
-                (goalkeeper == other.getGoalkeeper()) &&
-                (foul_counter == other.getFoulCounter()) &&
-                (ball_placement_failures == other.getBallPlacementFailures()) &&
-                (can_place_ball == other.getCanPlaceBall()) &&
-                (max_allowed_bots == other.getMaxAllowedBots()));
-    }
-
-   private:
-    // The team's name (empty string if operator has not typed anything).
-    std::string name;
-    // The number of goals scored by the team during normal play and overtime.
-    int score;
-    // The number of red cards issued to the team since the beginning of the game.
-    int red_cards;
-    // The amount of time (in microseconds) left on each yellow card issued to the team.
-    // If no yellow cards are issued, this array has no elements.
-    // Otherwise, times are ordered from smallest to largest.
-    std::vector<int> yellow_card_times;
-    // The total number of yellow cards ever issued to the team.
-    int yellow_cards;
-    // The number of timeouts this team can still call.
-    // If in a timeout right now, that timeout is excluded.
-    int timeouts;
-    // The number of microseconds of timeout this team can use.
-    int timeout_time;
-    // The pattern number of this team's goalkeeper.
-    int goalkeeper;
-    // The total number of countable fouls that act towards yellow cards
-    int foul_counter;
-    // The number of consecutive ball placement failures of this team
-    int ball_placement_failures;
-    // Indicate if the team is able and allowed to place the ball
-    bool can_place_ball;
-    // The maximum number of bots allowed on the field based on division and cards
-    int max_allowed_bots;
-};
-
 /**
  * RefboxData is a container class to represent the referee proto internally
  * It renames fields for ease of use, i.e. command->game_state and blue/yellow->us/them
@@ -293,93 +85,31 @@ class RefboxData
                TeamInfo enemy_team_info, RefboxGameState game_state,
                RefboxGameState next_game_state, RefboxStage stage,
                std::vector<GameEvent> game_events,
-               std::vector<ProposedGameEvent> proposed_game_events)
-        : packet_timestamp(packet_timestamp),
-          game_state_timestamp(game_state_timestamp),
-          game_state_counter(game_state_counter),
-          designated_position(designated_position),
-          blue_team_on_positive_half(blue_team_on_positive_half),
-          current_game_state_time_remaining(current_game_state_time_remaining),
-          friendly_team_info(friendly_team_info),
-          enemy_team_info(enemy_team_info),
-          game_state(game_state),
-          next_game_state(next_game_state),
-          stage(stage),
-          game_events(game_events),
-          proposed_game_events(proposed_game_events)
-    {
-    }
+               std::vector<ProposedGameEvent> proposed_game_events);
 
-    Timestamp getPacketTimestamp(void) const
-    {
-        return packet_timestamp;
-    }
+    Timestamp getPacketTimestamp(void) const;
 
-    Timestamp getGameStateTimestamp(void) const
-    {
-        return game_state_timestamp;
-    }
+    Timestamp getGameStateTimestamp(void) const;
 
-    int getGameStateCounter(void) const
-    {
-        return game_state_counter;
-    }
+    int getGameStateCounter(void) const;
 
-    Point getDesignatedPosition(void) const
-    {
-        return designated_position;
-    }
+    Point getDesignatedPosition(void) const;
 
-    bool getBlueTeamOnPositiveHalf(void) const
-    {
-        return blue_team_on_positive_half;
-    }
+    bool getBlueTeamOnPositiveHalf(void) const;
 
-    Duration getCurrentGameStateTimeRemaining(void) const
-    {
-        return current_game_state_time_remaining;
-    }
+    Duration getCurrentGameStateTimeRemaining(void) const;
 
-    TeamInfo getFriendlyTeamInfo(void) const
-    {
-        return friendly_team_info;
-    }
+    TeamInfo getFriendlyTeamInfo(void) const;
 
-    TeamInfo getEnemyTeamInfo(void) const
-    {
-        return enemy_team_info;
-    }
+    TeamInfo getEnemyTeamInfo(void) const;
 
-    RefboxGameState getGameState(void) const
-    {
-        return game_state;
-    }
+    RefboxGameState getGameState(void) const;
 
-    RefboxGameState getNextGameState(void) const
-    {
-        return next_game_state;
-    }
+    RefboxGameState getNextGameState(void) const;
 
-    RefboxStage getStage(void) const
-    {
-        return stage;
-    }
+    RefboxStage getStage(void) const;
 
-    bool operator==(const RefboxData &other) const
-    {
-        return ((packet_timestamp == other.getPacketTimestamp()) &&
-                (game_state_timestamp == other.getGameStateTimestamp()) &&
-                (game_state_counter == other.getGameStateCounter()) &&
-                (designated_position == other.getDesignatedPosition()) &&
-                (blue_team_on_positive_half == other.getBlueTeamOnPositiveHalf()) &&
-                (current_game_state_time_remaining ==
-                 other.getCurrentGameStateTimeRemaining()) &&
-                (friendly_team_info == other.getFriendlyTeamInfo()) &&
-                (enemy_team_info == other.getEnemyTeamInfo()) &&
-                (game_state == other.getGameState()) &&
-                (next_game_state == other.getNextGameState()) &&
-                (stage == other.getStage()));
-    }
+    bool operator==(const RefboxData &other) const;
 
    private:
     // packet timestamp
